refactor: refactor code
사용하지 않는 모듈을 제거합니다.
user모듈에 사용되는 dto를 dto 폴더 안으로 모읍니다.
JwtStrategy의 이름을 PasswordResetJwtStrategy로 변경합니다. 이는 auth모듈의 JwtStrategy와 헷갈리지 않게 하기 위함입니다.
