+++ backend/src/user/user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { JwtModule, JwtModuleOptions } from '@nestjs/jwt'
 import { EmailModule } from 'src/email/email.module'
-import { JwtStrategy } from './strategy/jwt-password-reset.strategy'
+import { PasswordResetJwtStrategy } from './strategy/jwt-password-reset.strategy'
 import { UserController } from './user.controller'
 import { UserService } from './user.service'
 
@@ -25,7 +25,7 @@ import { UserService } from './user.service'
     })
   ],
   controllers: [UserController],
-  providers: [UserService, JwtStrategy],
+  providers: [UserService, PasswordResetJwtStrategy],
   exports: [UserService]
 })
 export class UserModule {}
